+++ src/manage/denyAccessRequest.test.ts
@@ -58,13 +58,9 @@ describe("denyAccessRequest", () => {
     // eslint-disable-next-line @typescript-eslint/no-explicit-any
     const scab = jest.requireMock("@inrupt/solid-client-authn-browser") as any;
 
-    await denyAccessRequest(
-      "https://some.resource/owner",
-      mockAccessRequestVc()
-    );
+    await denyAccessRequest(mockAccessRequestVc());
 
     expect(mockedIssue).toHaveBeenCalledWith(
-      expect.anything(),
       expect.anything(),
       expect.anything(),
       expect.anything(),
@@ -77,7 +73,7 @@ describe("denyAccessRequest", () => {
   it("throws if the provided VC isn't a Solid access request", async () => {
     mockAccessApiEndpoint();
     await expect(
-      denyAccessRequest("https://some.resource/owner", {
+      denyAccessRequest({
         ...mockAccessRequestVc(),
         type: ["NotASolidAccessRequest"],
       })
@@ -88,9 +84,7 @@ describe("denyAccessRequest", () => {
 
   it("throws if there is no well known access endpoint", async () => {
     mockAccessApiEndpoint(false);
-    await expect(
-      denyAccessRequest("https://some.resource/owner", mockAccessRequestVc())
-    ).rejects.toThrow(
+    await expect(denyAccessRequest(mockAccessRequestVc())).rejects.toThrow(
       "No access issuer listed for property [verifiable_credential_issuer] in"
     );
   });
@@ -103,17 +97,12 @@ describe("denyAccessRequest", () => {
       mockedVcModule,
       "issueVerifiableCredential"
     );
-    await denyAccessRequest(
-      "https://some.resource/owner",
-      mockAccessRequestVc(),
-      {
-        accessEndpoint: "https://some.access-endpoint.override/",
-        fetch: jest.fn(),
-      }
-    );
+    await denyAccessRequest(mockAccessRequestVc(), {
+      accessEndpoint: "https://some.access-endpoint.override/",
+      fetch: jest.fn(),
+    });
     expect(spiedIssueRequest).toHaveBeenCalledWith(
       "https://some.access-endpoint.override/".concat("issue"),
-      expect.anything(),
       expect.anything(),
       expect.anything(),
       expect.anything()
@@ -141,7 +130,6 @@ describe("denyAccessRequest", () => {
       expect.anything(),
       expect.anything(),
       expect.anything(),
-      expect.anything(),
       { fetch: mockedFetch }
     );
   });
@@ -155,18 +143,13 @@ describe("denyAccessRequest", () => {
       mockedVcModule,
       "issueVerifiableCredential"
     );
-    await denyAccessRequest(
-      "https://some.resource/owner",
-      mockAccessRequestVc(),
-      {
-        fetch: jest.fn(global.fetch),
-      }
-    );
+    await denyAccessRequest(mockAccessRequestVc(), {
+      fetch: jest.fn(global.fetch),
+    });
 
     // TODO: Should we expect "isProvidedTo": "https://some.requestor" in "providedConsent" or nest the expect.objectContaining?
     expect(spiedIssueRequest).toHaveBeenCalledWith(
       `${MOCKED_ACCESS_ISSUER}/issue`,
-      "https://some.resource/owner",
       expect.objectContaining({
         providedConsent: {
           mode: mockAccessRequestVc().credentialSubject.hasConsent.mode,
@@ -198,19 +181,13 @@ describe("denyAccessRequest", () => {
       .mockResolvedValueOnce(
         new Response(JSON.stringify(mockAccessRequestVc()))
       );
-    await denyAccessRequest(
-      "https://some.resource/owner",
-      "https://some.credential",
-      {
-        fetch: mockedFetch,
-      }
-    );
+    await denyAccessRequest("https://some.credential", {
+      fetch: mockedFetch,
+    });
 
     expect(spiedIssueRequest).toHaveBeenCalledWith(
       `${MOCKED_ACCESS_ISSUER}/issue`,
-      "https://some.resource/owner",
       expect.objectContaining({
-        id: "https://some.resource/owner",
         providedConsent: expect.objectContaining({
           mode: mockAccessRequestVc().credentialSubject.hasConsent.mode,
           hasStatus: "https://w3id.org/GConsent#ConsentStatusDenied",
@@ -240,26 +217,95 @@ describe("denyAccessRequest", () => {
       .mockResolvedValueOnce(
         new Response(JSON.stringify(mockAccessRequestVc()))
       );
-    await denyAccessRequest(
-      "https://some.resource/owner",
-      new URL("https://some.credential"),
-      {
-        fetch: mockedFetch,
-      }
-    );
+    await denyAccessRequest(new URL("https://some.credential"), {
+      fetch: mockedFetch,
+    });
 
     // TODO: Should we expect "isProvidedTo": "https://some.requestor" in "providedConsent" or nest the expect.objectContaining?
     expect(spiedIssueRequest).toHaveBeenCalledWith(
       `${MOCKED_ACCESS_ISSUER}/issue`,
-      "https://some.resource/owner",
       expect.objectContaining({
-        id: "https://some.resource/owner",
         providedConsent: expect.objectContaining({
           mode: mockAccessRequestVc().credentialSubject.hasConsent.mode,
           hasStatus: "https://w3id.org/GConsent#ConsentStatusDenied",
           forPersonalData:
             mockAccessRequestVc().credentialSubject.hasConsent.forPersonalData,
         }),
+        inbox: mockAccessRequestVc().credentialSubject.inbox,
+      }),
+      expect.objectContaining({
+        type: ["SolidAccessDenial"],
+      }),
+      expect.anything()
+    );
+  });
+
+  it("issues a proper denied access VC using the deprecated signature and VC value", async () => {
+    mockAccessApiEndpoint();
+    const mockedVcModule = jest.requireMock("@inrupt/solid-client-vc") as {
+      issueVerifiableCredential: () => unknown;
+    };
+    const spiedIssueRequest = jest.spyOn(
+      mockedVcModule,
+      "issueVerifiableCredential"
+    );
+    await denyAccessRequest(
+      "https://some.resource.owner",
+      mockAccessRequestVc(),
+      {
+        fetch: jest.fn(global.fetch),
+      }
+    );
+
+    // TODO: Should we expect "isProvidedTo": "https://some.requestor" in "providedConsent" or nest the expect.objectContaining?
+    expect(spiedIssueRequest).toHaveBeenCalledWith(
+      `${MOCKED_ACCESS_ISSUER}/issue`,
+      expect.objectContaining({
+        providedConsent: {
+          mode: mockAccessRequestVc().credentialSubject.hasConsent.mode,
+          hasStatus: "https://w3id.org/GConsent#ConsentStatusDenied",
+          forPersonalData:
+            mockAccessRequestVc().credentialSubject.hasConsent.forPersonalData,
+          isProvidedTo: "https://some.requestor",
+        },
+        inbox: mockAccessRequestVc().credentialSubject.inbox,
+      }),
+      expect.objectContaining({
+        type: ["SolidAccessDenial"],
+      }),
+      expect.anything()
+    );
+  });
+
+  it("issues a proper denied access VC using the deprecated signature and VC IRI", async () => {
+    mockAccessApiEndpoint();
+    const mockedVcModule = jest.requireMock("@inrupt/solid-client-vc") as {
+      issueVerifiableCredential: () => unknown;
+    };
+    const spiedIssueRequest = jest.spyOn(
+      mockedVcModule,
+      "issueVerifiableCredential"
+    );
+    // eslint-disable-next-line @typescript-eslint/no-explicit-any
+    const scab = jest.requireMock("@inrupt/solid-client-authn-browser") as any;
+    scab.fetch = jest
+      .fn(global.fetch)
+      .mockResolvedValueOnce(
+        new Response(JSON.stringify(mockAccessRequestVc()))
+      );
+    await denyAccessRequest("https://some.resource.owner", "https://some.vc");
+
+    // TODO: Should we expect "isProvidedTo": "https://some.requestor" in "providedConsent" or nest the expect.objectContaining?
+    expect(spiedIssueRequest).toHaveBeenCalledWith(
+      `${MOCKED_ACCESS_ISSUER}/issue`,
+      expect.objectContaining({
+        providedConsent: {
+          mode: mockAccessRequestVc().credentialSubject.hasConsent.mode,
+          hasStatus: "https://w3id.org/GConsent#ConsentStatusDenied",
+          forPersonalData:
+            mockAccessRequestVc().credentialSubject.hasConsent.forPersonalData,
+          isProvidedTo: "https://some.requestor",
+        },
         inbox: mockAccessRequestVc().credentialSubject.inbox,
       }),
       expect.objectContaining({